refactor: relocate ritual and routine models to core module
This commit updates the nutrition feature to use the centralized ritual and routine models. These models have been moved from the `kocolor` feature package to `com.zoewave.probase.core.model.ritual` to improve modularity and support shared usage across different features.

- **`features:health:nutrition`**:
    - Updated imports for `BeautyRoutine`, `RoutineStep`, and `RoutineTime` in `NutritionRitualStep.kt`, `NutritionUiRoute.kt`, `NutritionUiState.kt`, and `NutritionViewModel.kt`.
    - Pointed ritual-related data structures to the new `com.zoewave.probase.core.model.ritual` package.
