+++ features/health/nutrition/src/main/java/com/zoewave/probase/features/health/nutrition/ui/NutritionUiRoute.kt
@@ -21,8 +21,8 @@ import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.zoewave.probase.features.health.nutrition.ui.components.NutritionRitualHeader
 import com.zoewave.probase.features.health.nutrition.ui.components.NutritionRitualStep
-import com.zoewave.probase.kocolor.model.BeautyRoutine
-import com.zoewave.probase.kocolor.model.RoutineStep
+import com.zoewave.probase.core.model.ritual.BeautyRoutine
+import com.zoewave.probase.core.model.ritual.RoutineStep
 import com.zoewave.probase.kocolor.model.KoColorRoute
 
 @Composable
